Refactor TempDir from TileDBTestUtils into catch.h to use in everywhere. Otherwise, some tests are failing on NFS
