+++ test/include/catch2/catch.h
@@ -1,5 +1,7 @@
-#define CATCH_CONFIG_MAIN
+#define CATCH_CONFIG_RUNNER
 #include "catch.hpp"
+
+#include "tiledb_utils.h"
 
 // Matchers
 using Catch::Equals;
@@ -9,3 +11,86 @@ using Catch::Contains;
 using Catch::Matches;
 
 #define CHECK_RC(rc, expected) CHECK(rc == expected)
+
+std::string g_test_dir = "";
+
+class TempDir {
+ public:
+  TempDir() {
+    create_temp_directory();
+  }
+
+  ~TempDir() {
+    TileDBUtils::delete_dir(get_temp_dir());
+
+    if (!delete_test_dir_in_destructor_.empty()) {
+      TileDBUtils::delete_dir(delete_test_dir_in_destructor_);
+    }
+  }
+
+  const std::string& get_temp_dir() {
+    return tmp_dirname_;
+  }
+ private:
+  std::string tmp_dirname_;
+  std::string delete_test_dir_in_destructor_;
+
+  std::string get_pathname(std::string  path) {
+    const size_t last_slash_idx = path.rfind('/');
+    if (last_slash_idx != std::string::npos) {
+      return path.substr(last_slash_idx+1);
+    } else {
+      return path;
+    }
+  }
+  std::string append_slash(std::string path) {
+    if (path[path.size()]!='/') {
+      return path+"/";
+    } else {
+      return path;
+    }
+  }
+  void create_temp_directory() {
+    std::string dirname_pattern("TileDBTestXXXXXX");
+    if (g_test_dir.empty()) { // Posix Case. Use mkdtemp() here
+      const char *tmp_dir = getenv("TMPDIR");
+      if (tmp_dir == NULL) {
+        tmp_dir = P_tmpdir; // defined in stdio
+      }
+      assert(tmp_dir != NULL);
+      tmp_dirname_ = mkdtemp(const_cast<char *>((append_slash(tmp_dir)+dirname_pattern).c_str()));
+    } else {
+      tmp_dirname_ = append_slash(g_test_dir)+mktemp(const_cast<char *>(dirname_pattern.c_str()));
+      if (!TileDBUtils::is_dir(g_test_dir)) {
+        CHECK(TileDBUtils::create_dir(g_test_dir) == 0);
+        delete_test_dir_in_destructor_ = g_test_dir;
+      }
+      if (!TileDBUtils::is_dir(tmp_dirname_)) {
+        CHECK(TileDBUtils::create_dir(tmp_dirname_) == TILEDB_OK);
+      }
+    }
+  }
+};
+
+int main( int argc, char* argv[] )
+{
+  Catch::Session session; // There must be exactly one instance
+
+  // Build a new parser on top of Catch's
+  using namespace Catch::clara;
+  auto cli
+    = session.cli() // Get Catch's composite command line parser
+    | Opt( g_test_dir, "Specify test dir, default is $TMPDIR" ) // bind variable to a new option, with a hint string
+     ["--test-dir"] // the option names it will respond to
+    ("Specify test dir, default is $TMPDIR if not specified");        // description string for the help output
+
+  // Now pass the new composite back to Catch so it uses that
+  session.cli(cli);
+
+  // Let Catch (using Clara) parse the command line
+  int rc = session.applyCommandLine( argc, argv );
+  if(rc != 0) // Indicates a command line error
+    return rc;
+
+  return session.run();
+}
